// src/modules/profile/models/profile-colors.ts
export enum ProfileColors {
  CountdownText = "countdownTextColor",
  CountdownTitle = "countdownTitleColor",
  CountdownBg = "countdownBackgroundColor",
  NavbarText = "navbarTextColor",
  NavbarBg = "navbarBackgroundColor",
  TabbarText = "tabbarTextColor",
  TabbarUnselected = "tabbarUnselectedColor",
  TabbarBg = "tabbarBackgroundColor",
  InvintationBg = "invitationBackgroundColor",
  InvintationTitle = "invitationTitleColor",
  InvintationText = "invitationTextColor",
  InvintationButton = "invitationButtonColor",
}